requirements.txt
streamlit>=1.30
pandas
yfinance
pyyaml
requests
pytz
ta
plotly
streamlit-aggrid
streamlit-autorefresh
plyer